text-to-image-generator/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers and Diffusion
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Scientific computing
numpy>=1.26.0,<2.0.0
scipy>=1.11.0
matplotlib>=3.7.0
scikit-learn>=1.3.0

# Web framework
Flask>=2.3.0
Flask-Cors>=4.0.0

# Utilities
tqdm>=4.65.0
huggingface-hub>=0.16.0
safetensors>=0.3.0

# Optional but recommended
xformers>=0.0.20  # For memory-efficient attention
omegaconf>=2.3.0  # For configuration management